Folder structure updated, api services refactored.
