+++ frontend/src/App.js
@@ -2,8 +2,8 @@ import React from "react";
 import { Routes, Route } from "react-router-dom";
 import './App.css';
 import Home from "./home_page/Home";
-import Login from "./login_page/Login";
-import useToken from './useToken';
+import LoginPage from "./login_page/LoginPage";
+import useToken from './services/useToken';
 
 /**
  * The App function, where users can navigate to and from different pages
@@ -15,13 +15,13 @@ function App() {
   const { token, setToken } = useToken();
 
   if (!token) { // TODO: check if session token is also valid via http request to database
-    return <Login setToken={setToken}/>
+    return <LoginPage setToken={setToken}/>
   }
 
   return (
     <div className="App">
       <Routes>
-        <Route path="/" element={<Login setToken={setToken}/>}/>
+        <Route path="/" element={<LoginPage setToken={setToken}/>}/>
         <Route path="/home" element={<Home/>}/>
       </Routes>
     </div>
