+++ frontend/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom'
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
+import reportWebVitals from './services/reportWebVitals';
 
 /**
  * Where the App function is called and the website instance is rendered
